user_db integration
integrated user_db into the larger project
