+++ user_profile_db/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Profile, Course
+# from profilepage.models import Profile, Course
 
-admin.site.register(Profile)
-admin.site.register(Course)
+# admin.site.register(Profile)
+# admin.site.register(Course)
