+++ user_profile_db/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 
-from .models import Course, Profile
+from profilepage.models import Course, Profile
 from .forms import add_course_form, remove_course_form
 
 # Create your views here.
